# docs/api/booking-v1.openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail"
          }
        },
        "required": [
          "code",
          "detail"
        ],
        "title": "ApiError",
        "type": "object"
      },
      "Appointment": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "duration_min": {
            "title": "Duration Min",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "local_date": {
            "title": "Local Date",
            "type": "string"
          },
          "local_time": {
            "title": "Local Time",
            "type": "string"
          },
          "patient_id": {
            "title": "Patient Id",
            "type": "integer"
          },
          "patient_name": {
            "title": "Patient Name",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "start_at": {
            "title": "Start At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "therapist_id": {
            "title": "Therapist Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "therapist_id",
          "patient_id",
          "patient_name",
          "start_at",
          "local_date",
          "local_time",
          "duration_min",
          "status",
          "source",
          "created_at"
        ],
        "title": "Appointment",
        "type": "object"
      },
      "AppointmentList": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Appointment"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "AppointmentList",
        "type": "object"
      },
      "BookingIn": {
        "properties": {
          "duration_min": {
            "default": 60,
            "maximum": 60.0,
            "minimum": 60.0,
            "title": "Duration Min",
            "type": "integer"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientIn"
          },
          "send_confirmation": {
            "default": true,
            "title": "Send Confirmation",
            "type": "boolean"
          },
          "source": {
            "default": "api",
            "enum": [
              "api",
              "telegram",
              "whatsapp",
              "manual"
            ],
            "title": "Source",
            "type": "string"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          },
          "summary": {
            "default": "",
            "maxLength": 2000,
            "title": "Summary",
            "type": "string"
          },
          "therapist_id": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Therapist Id",
            "type": "string"
          }
        },
        "required": [
          "therapist_id",
          "start_at",
          "patient"
        ],
        "title": "BookingIn",
        "type": "object"
      },
      "PatientIn": {
        "properties": {
          "channel": {
            "anyOf": [
              {
                "enum": [
                  "telegram",
                  "whatsapp"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "email": {
            "anyOf": [
              {
                "maxLength": 254,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "external_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "patient_id": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "phone": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "required": [
          "name"
        ],
        "title": "PatientIn",
        "type": "object"
      },
      "Slot": {
        "properties": {
          "duration_min": {
            "title": "Duration Min",
            "type": "integer"
          },
          "local_date": {
            "title": "Local Date",
            "type": "string"
          },
          "local_time": {
            "title": "Local Time",
            "type": "string"
          },
          "start_at": {
            "title": "Start At",
            "type": "string"
          }
        },
        "required": [
          "start_at",
          "local_date",
          "local_time",
          "duration_min"
        ],
        "title": "Slot",
        "type": "object"
      },
      "SlotList": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Slot"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "SlotList",
        "type": "object"
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "description": "A machine client's API key (`python -m zenflow.api_keys create`). The dashboard uses its session cookie instead.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Booking for ZenFlow Clinic. Every appointment in the system is created through `POST /api/v1/appointments`.",
    "title": "ZenFlow Booking API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/appointments": {
      "get": {
        "operationId": "list_appointments_api_v1_appointments_get",
        "parameters": [
          {
            "in": "query",
            "name": "therapist_id",
            "required": true,
            "schema": {
              "maxLength": 32,
              "minLength": 1,
              "title": "Therapist Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "active",
              "enum": [
                "active",
                "cancelled",
                "all"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "patient_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentList"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No or bad credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not this client's therapist"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unknown therapist, patient or appointment"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The request does not make sense"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Too many requests"
          }
        },
        "summary": "List appointments",
        "tags": [
          "booking"
        ]
      },
      "post": {
        "operationId": "create_appointment_api_v1_appointments_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No or bad credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not this client's therapist"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unknown therapist, patient or appointment"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The hour is gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The request does not make sense"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Too many requests"
          }
        },
        "summary": "Book an appointment",
        "tags": [
          "booking"
        ]
      }
    },
    "/api/v1/appointments/{appointment_id}": {
      "delete": {
        "operationId": "cancel_appointment_api_v1_appointments__appointment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "appointment_id",
            "required": true,
            "schema": {
              "title": "Appointment Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No or bad credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not this client's therapist"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unknown therapist, patient or appointment"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The request does not make sense"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Too many requests"
          }
        },
        "summary": "Cancel an appointment",
        "tags": [
          "booking"
        ]
      },
      "get": {
        "operationId": "get_appointment_api_v1_appointments__appointment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "appointment_id",
            "required": true,
            "schema": {
              "title": "Appointment Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No or bad credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not this client's therapist"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unknown therapist, patient or appointment"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The request does not make sense"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Too many requests"
          }
        },
        "summary": "One appointment",
        "tags": [
          "booking"
        ]
      }
    },
    "/api/v1/availability": {
      "get": {
        "operationId": "get_availability_api_v1_availability_get",
        "parameters": [
          {
            "in": "query",
            "name": "therapist_id",
            "required": true,
            "schema": {
              "maxLength": 32,
              "minLength": 1,
              "title": "Therapist Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "title": "To",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlotList"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No or bad credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not this client's therapist"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unknown therapist, patient or appointment"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The request does not make sense"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Too many requests"
          }
        },
        "summary": "Free hours",
        "tags": [
          "booking"
        ]
      }
    }
  },
  "security": [
    {
      "ApiKey": []
    }
  ]
}